refactor: replace motion with view transition for product image animation
